Reorganize project directory structure
