+++ src/new-frontend/src/components/UserSettings/DeleteAccount.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 
 import { Button, Container, Heading, Text, useDisclosure } from '@chakra-ui/react';
 
-import DeleteConfirmation from '../modals/DeleteConfirmation';
+import DeleteConfirmation from './DeleteConfirmation';
 
 const DeleteAccount: React.FC = () => {
     const confirmationModal = useDisclosure();
