+++ src/new-frontend/src/components/UserSettings/DeleteConfirmation.tsx
